+++ cloud_agent/__init__.py
@@ -1,5 +1,3 @@
-# /home/jono/ai_orchestrator/cloud_agent/__init__.py
-from .cloud_client import client, get_plan
-__all__ = ["client", "get_plan"]
+from .cloud_client import get_plan
 
-
+__all__ = ["get_plan"]